@@ -1,21 +1,21 @@
-const express = require('express');
-const routes = require('./routes');
+import express from 'express';
+import routes from './routes';
 
-class App{
-  constructor(){
+class App {
+  constructor() {
     this.server = express();
 
     this.middlewares();
     this.routes();
   }
 
-  middlewares(){
+  middlewares() {
     this.server.use(express.json());
   }
 
-  routes(){
+  routes() {
     this.server.use(routes);
   }
 }
 
-module.exports = new App().server;
+export default new App().server;
