@@ -1,10 +1,7 @@
-const {Router} = require('express');
+import { Router } from 'express';
 
 const routes = new Router();
 
-routes.get('/', (req, res) => {
-  return res.json({a: "olar FastFeed"});
-});
+routes.get('/', (req, res) => res.json({ main: 'Hello FastFeet' }));
 
-
-module.exports = routes;
+export default routes;
